Start to make things unit-testable
